+++ packages/microsoft-reactnative-sampleapps/windows/SampleLibraryCPP/CustomUserControlViewManagerCPP.h
@@ -7,16 +7,20 @@
 
 namespace winrt::SampleLibraryCPP::implementation {
 
-struct CustomUserControlViewManagerCPP : winrt::implements<
-                                             CustomUserControlViewManagerCPP,
-                                             winrt::Microsoft::ReactNative::Bridge::IViewManager,
-                                             winrt::Microsoft::ReactNative::Bridge::IViewManagerWithNativeProperties,
-                                             winrt::Microsoft::ReactNative::Bridge::IViewManagerWithCommands> {
+struct CustomUserControlViewManagerCPP
+    : winrt::implements<
+          CustomUserControlViewManagerCPP,
+          winrt::Microsoft::ReactNative::Bridge::IViewManager,
+          winrt::Microsoft::ReactNative::Bridge::IViewManagerWithNativeProperties,
+          winrt::Microsoft::ReactNative::Bridge::IViewManagerWithCommands,
+          winrt::Microsoft::ReactNative::Bridge::IViewManagerWithExportedEventTypeConstants> {
  public:
-  CustomUserControlViewManagerCPP() = default;
+  CustomUserControlViewManagerCPP(winrt::Microsoft::ReactNative::Bridge::IReactContext const &reactContext);
 
   // IViewManager
   winrt::hstring Name() noexcept;
+
+  winrt::Microsoft::ReactNative::Bridge::IReactContext ReactContext() noexcept;
 
   winrt::Windows::UI::Xaml::FrameworkElement CreateView() noexcept;
 
@@ -38,6 +42,16 @@ struct CustomUserControlViewManagerCPP : winrt::implements<
       int64_t commandId,
       winrt::Windows::Foundation::Collections::IVectorView<winrt::Windows::Foundation::IInspectable>
           commandArgs) noexcept;
+
+  // IViewManagerWithExportedEventTypeConstants
+  winrt::Windows::Foundation::Collections::IMapView<winrt::hstring, winrt::Windows::Foundation::IInspectable>
+  ExportedCustomBubblingEventTypeConstants() noexcept;
+
+  winrt::Windows::Foundation::Collections::IMapView<winrt::hstring, winrt::Windows::Foundation::IInspectable>
+  ExportedCustomDirectEventTypeConstants() noexcept;
+
+ private:
+  winrt::Microsoft::ReactNative::Bridge::IReactContext m_reactContext{nullptr};
 };
 
 } // namespace winrt::SampleLibraryCPP::implementation