+++ packages/microsoft-reactnative-sampleapps/windows/SampleLibraryCPP/ReactPackageProvider.cpp
@@ -7,6 +7,7 @@
 #include "ReactPackageProvider.g.cpp"
 #endif
 
+#include "CircleViewManagerCPP.h"
 #include "CustomUserControlViewManagerCPP.h"
 #include "SampleModuleCPP.h"
 
@@ -17,8 +18,11 @@ namespace winrt::SampleLibraryCPP::implementation {
 
 void ReactPackageProvider::CreatePackage(IReactPackageBuilder const &packageBuilder) noexcept {
   AddAttributedModules(packageBuilder);
+  packageBuilder.AddViewManager(L"CustomUserControlViewManagerCPP", [](IReactContext const &reactContext) {
+    return winrt::make<CustomUserControlViewManagerCPP>(reactContext);
+  });
   packageBuilder.AddViewManager(
-      L"CustomUserControlViewManagerCPP", []() { return winrt::make<CustomUserControlViewManagerCPP>(); });
+      L"CircleViewManagerCPP", [](IReactContext const &reactContext) { return winrt::make<CircleViewManagerCPP>(); });
 }
 
 } // namespace winrt::SampleLibraryCPP::implementation