+++ vnext/Microsoft.ReactNative/ABIViewManager.h
@@ -43,6 +43,21 @@ class ABIViewManager : public react::uwp::FrameworkElementViewManager {
       int64_t commandId,
       const folly::dynamic &commandArgs) override;
 
+  folly::dynamic GetExportedCustomBubblingEventTypeConstants() const override;
+
+  folly::dynamic GetExportedCustomDirectEventTypeConstants() const override;
+
+  void AddView(
+      winrt::Windows::UI::Xaml::DependencyObject parent,
+      winrt::Windows::UI::Xaml::DependencyObject child,
+      int64_t index) override;
+  void RemoveAllChildren(winrt::Windows::UI::Xaml::DependencyObject parent) override;
+  void RemoveChildAt(winrt::Windows::UI::Xaml::DependencyObject parent, int64_t index) override;
+  void ReplaceChild(
+      winrt::Windows::UI::Xaml::DependencyObject parent,
+      winrt::Windows::UI::Xaml::DependencyObject oldChild,
+      winrt::Windows::UI::Xaml::DependencyObject newChild) override;
+
  protected:
   winrt::Windows::UI::Xaml::DependencyObject CreateViewCore(int64_t) override;
 
@@ -51,6 +66,9 @@ class ABIViewManager : public react::uwp::FrameworkElementViewManager {
   winrt::Microsoft::ReactNative::Bridge::IViewManagerWithExportedViewConstants m_viewManagerWithExportedViewConstants;
   winrt::Microsoft::ReactNative::Bridge::IViewManagerWithNativeProperties m_viewManagerWithNativeProperties;
   winrt::Microsoft::ReactNative::Bridge::IViewManagerWithCommands m_viewManagerWithCommands;
+  winrt::Microsoft::ReactNative::Bridge::IViewManagerWithExportedEventTypeConstants
+      m_viewManagerWithExportedEventTypeConstants;
+  winrt::Microsoft::ReactNative::Bridge::IViewManagerWithChildren m_viewManagerWithChildren;
 
   winrt::Windows::Foundation::Collections::IMapView<winrt::hstring, ViewManagerPropertyType> m_nativeProps;
 };