+++ vnext/Microsoft.ReactNative/ViewManagersProvider.cpp
@@ -4,6 +4,8 @@
 #include "pch.h"
 #include "ViewManagersProvider.h"
 
+#include "IReactContext.h"
+
 #include "ABIViewManager.h"
 
 namespace winrt::Microsoft::ReactNative::Bridge {
@@ -17,7 +19,8 @@ std::vector<react::uwp::NativeViewManager> ViewManagersProvider::GetViewManagers
   for (auto &entry : m_viewManagerProviders) {
     auto viewManagerProvider = entry.second;
 
-    auto viewManager = std::make_unique<ABIViewManager>(instance, viewManagerProvider());
+    auto reactContext = winrt::make<ReactContext>(instance).as<IReactContext>();
+    auto viewManager = std::make_unique<ABIViewManager>(instance, viewManagerProvider(reactContext));
 
     viewManagers.emplace_back(std::move(viewManager));
   }